Refactor memory layer; Refactor AVR executor to not automatically modify PC on PC-modifying instructions; Add BR, IN, OUT instructions;

+++ plugins/architecture_avr/Instructions/Call/CallInstruction.cs
@@ -13,8 +13,8 @@ namespace architecture_avr.Instructions.Call
 
         public override void Execute(AvrCpuState cpuState, DeviceEnvironment env)
         {
-            env.MemoryMap.WriteByte(cpuState.Sp--, (byte)(cpuState.Pc + 2));
-            env.MemoryMap.WriteByte(cpuState.Sp--, (byte)((cpuState.Pc + 2) >> 8));
+            env.MemoryMap.WriteByte((int)cpuState.Sp--, (byte)(cpuState.Pc + 2));
+            env.MemoryMap.WriteByte((int)cpuState.Sp--, (byte)((cpuState.Pc + 2) >> 8));
 
             cpuState.Pc = (uint)_immAdd;
         }
